AI/api/requirements.txt
fastapi
uvicorn
keras==3.1.0
tensorflow==2.16.1
numpy
pillow
requests
python-multipart
